made code not stupid
