+++ PaintingPack.json
@@ -1,1 +1,1 @@
-{"bust": {"detail": 16, "frameName": "bust", "size": "2x2", "scale_method": "Stretch", "background_color": "#000000", "frame_index": 1, "file_path": "/home/rory/Desktop/input/Bust.jpg"}, "kebab": {"detail": 1, "frameName": "kebab", "size": "1x1", "scale_method": "Stretch", "background_color": "#000000", "frame_index": 0, "file_path": "/home/rory/Desktop/input/bomb.png"}}
+{"bust": {"detail": 16, "frameName": "bust", "size": "2x2", "scale_method": "Stretch", "background_color": "#000000", "frame_index": 1, "file_path": "/home/rory/Desktop/input/Bust.jpg"}}
