+++ PaintingStudio.py
@@ -11,8 +11,8 @@ from io import BytesIO
 from PIL import Image
 from PaintingGenerator import PaintingGenerator
 from ResourcePackBuilder import ResourcePackBuilder
-from FrameDialog import LoadingDialog, InputDialog
-from FrameWidgets import PackControls
+from FrameDialog import LoadingDialog, InputDialog, HelpDialog
+from FrameWidgets import PackControls, PaintingEditor
 
 class PaintingStudio(QMainWindow):
 
@@ -22,19 +22,6 @@ class PaintingStudio(QMainWindow):
         self.setCentralWidget(self.central_widget)
         layout = QHBoxLayout(self)
         self.central_widget.setLayout(layout)
-        init_silder_value = 500
-        self.view_size = int(100 + (init_silder_value / 500) * 300)
-
-        with open(self.resource_path('paintings.json'), 'r') as file:
-            self.paintings = json.load(file)
-        self.file_path_stack = []
-        self.drawThread = QTimer(self)
-        self.drawThread.timeout.connect(self.forceViewPortDraw)
-        self.drawThread.setSingleShot(True)
-        self.lock = True
-        self.updating = False
-        self.packCreated = False
-        self.backgroundColor = "#000000"
 
         # generated stuff
         self.setWindowTitle("Minecraft Painting Studio")
@@ -61,133 +48,12 @@ class PaintingStudio(QMainWindow):
 
         """ Left Bar """
         self.packConrols = PackControls(self)
-
-
-        """Options Pane"""
-        PaintingOptions = QWidget()
-        PaintingOptions_layout = QVBoxLayout()
-
-        lable_width = 120
-        detail_layout = QHBoxLayout()
-        self.detail_spin_box = QSpinBox(self)
-        self.detail_spin_box.setRange(1, 16)  # Set the valid range (1 to 100)
-        self.detail_spin_box.setValue(1)  # Set the initial value
-        self.detail_spin_box.valueChanged.connect(self.requestViewPortDraw)
-        detail_label = QLabel('Detail: ')
-        detail_label.setMaximumWidth(lable_width)
-        detail_layout.addWidget(detail_label)
-        detail_layout.addWidget(self.detail_spin_box)
-
-        scale_layout = QHBoxLayout()
-        self.scale_combo_box = QComboBox(self)
-        scale_label = QLabel('Scale Method: ')
-        scale_label.setMaximumWidth(lable_width)
-        scale_layout.addWidget(scale_label)
-        self.scaleOptions = ["Stretch", "Fit", "Crop"]
-        self.scale_combo_box.addItems(self.scaleOptions)
-        scale_layout.addWidget(self.scale_combo_box)
-        color_button_layout = QHBoxLayout()
-        self.color_button = QPushButton('Choose Backgroud Color', self)
-        self.color_button.clicked.connect(self.showColorDialog)
-        color_button_layout.addWidget(self.color_button)
-
-        size_layout = QHBoxLayout()
-        self.size_combo_box = QComboBox(self)
-        size_label = QLabel('Painting Size: ')
-        size_label.setMaximumWidth(lable_width)
-        size_layout.addWidget(size_label)
-        size_layout.addWidget(self.size_combo_box)
-
-        painting_layout = QHBoxLayout()
-        self.painting_combo_box = QComboBox(self)
-        painting_label = QLabel('Painting: ')
-        painting_label.setMaximumWidth(lable_width)
-        painting_layout.addWidget(painting_label)
-        painting_layout.addWidget(self.painting_combo_box)
-        frame_layout = QHBoxLayout()
-        self.frame_combo_box = QComboBox(self)
-        frame_label = QLabel('Frame: ')
-        frame_label.setMaximumWidth(lable_width//2)
-        frame_layout.addWidget(frame_label)
-        frame_layout.addWidget(self.frame_combo_box)
-
-        for key in self.paintings:
-            self.size_combo_box.addItem(key)
-        self.updateComboBox()
-        self.size_combo_box.currentIndexChanged.connect(self.updateComboBox)
-        self.painting_combo_box.currentIndexChanged.connect(self.updateFrameComboBox)
-        self.scale_combo_box.currentIndexChanged.connect(self.requestViewPortDraw)
-        self.frame_combo_box.currentIndexChanged.connect(self.requestViewPortDraw)
-
-        # Add Layouts
-        PaintingOptions_layout.addWidget(QLabel("<br><b>Framing Options</b>"))
-        PaintingOptions_layout.addLayout(detail_layout)
-        PaintingOptions_layout.addLayout(scale_layout)
-        PaintingOptions_layout.addLayout(color_button_layout)
-        PaintingOptions_layout.addWidget(QLabel("<br><b>Painting Options</b>"))
-        PaintingOptions_layout.addLayout(size_layout)
-        PaintingOptions_layout.addLayout(painting_layout)
-        PaintingOptions_layout.addLayout(frame_layout)
-        PaintingOptions_layout.addStretch()
-        PaintingOptions.setLayout(PaintingOptions_layout)
-
-
-        """View Port"""
-        ViewPort = QWidget()
-        ViewPort_layout = QVBoxLayout()
-        # Main View Port
-        scroll_area = QScrollArea(self)
-        self.image_label = QLabel("Drop image here to customize your painting", self)
-        scroll_area.setWidget(self.image_label)
-        scroll_area.setAlignment(Qt.AlignCenter)
-        self.image_label.setAlignment(Qt.AlignCenter)
-        self.image_label.setWordWrap(True)
-        self.image_label.setFixedSize(self.view_size, self.view_size)
-        #Add Layouts
-        ViewPort_layout.addWidget(scroll_area)
-        ViewPort.setLayout(ViewPort_layout)
-
-
-        """Tool Bar"""
-        ToolBar = QWidget()
-        ToolBar_layout = QVBoxLayout()
-        # View Port Tools
-        tools_layout = QHBoxLayout()
-        self.path_label = QLabel(" ", self)
-        self.path_label.setMinimumWidth(1)
-        #self.path_label.setSizePolicy(QSizePolicy.Ignored, QSizePolicy.Fixed)
-        self.view_slider = QSlider(Qt.Horizontal)
-        self.view_slider.setRange(0,1000)  # Set minimum value
-        self.view_slider.setValue(init_silder_value)  # Set initial value
-        self.view_slider.setTickPosition(QSlider.TicksBelow)
-        self.view_slider.setTickInterval(125)
-        self.view_slider.setFixedWidth(100)
-        self.view_slider.setMaximumWidth(150)
-        self.view_slider.valueChanged.connect(self.view_slider_changed)
-        tools_layout.addWidget(self.path_label)
-        tools_layout.addStretch()
-        tools_layout.addWidget(self.view_slider)
-        #Add Layouts
-        ToolBar_layout.addLayout(tools_layout)
-        ToolBar.setLayout(ToolBar_layout)
-
+        
+        """Center Widget"""
+        self.paintingEditor = PaintingEditor(self)
 
         """Main Window"""
-
-        ViewPort.setMinimumWidth(600)
-        PaintingOptions.setMinimumWidth(250)
-        PaintingOptions.setMaximumWidth(350)
-        self.packConrols.setMinimumWidth(250)
-        self.packConrols.setMaximumWidth(350)
-
-        combine_OptionsViewport = QHBoxLayout()
-        combine_ViewportToolbar = QVBoxLayout()
-
-        combine_OptionsViewport.addWidget(PaintingOptions)
-        combine_OptionsViewport.addWidget(ViewPort)
-        combine_ViewportToolbar.addLayout(combine_OptionsViewport)
-        combine_ViewportToolbar.addWidget(ToolBar)
-        layout.addLayout(combine_ViewportToolbar)
+        layout.addWidget(self.paintingEditor)
         layout.addWidget(self.packConrols)
         self.setLayout(layout)
 
@@ -198,13 +64,11 @@ class PaintingStudio(QMainWindow):
     def newPack(self):
         # Create and show the input dialog
         dialog = InputDialog(self)
-
+        
         # Check if the dialog was accepted
         if dialog.exec_() == QDialog.Accepted:
             title, description, number, icon = dialog.get_data()
-            self.packCreated = True
-            self.painting_maker = PaintingGenerator()
-
+            self.paintingEditor.newPack()
             #remove self. later
             self.packName = title
             packMeta = {
@@ -215,41 +79,14 @@ class PaintingStudio(QMainWindow):
             }
             self.packConrols.setPackInfo(title, packMeta, icon)
 
-            self.size_combo_box.clear()
-            for key in self.paintings:
-                self.size_combo_box.addItem(key)
-            self.updateComboBox()
-
     def reset(self):
-        for key in self.paintings:
-            self.size_combo_box.addItem(key)
-        self.updateComboBox()
-        self.setButtonEnabled(False)
-        self.image_label.clear()
-        self.image_label.setText("Drop image here to customize your painting")
-        self.path_label.setText("")
-        self.view_size = 400
-        self.image_label.setFixedSize(self.view_size, self.view_size)
-
-
-    def view_slider_changed(self):
-        self.requestViewPortDraw()
-
-    def update_view_size(self, autoScale = False):
-        value = self.view_slider.value()
-        if value <= 500:
-            # Bottom half (100 to 400)
-            self.view_size = int(100 + (value / 500) * 300)
-        else:
-            # Top half (400 to 1600)
-            self.view_size = int(400 + ((value - 500) / 500) * 1200)
-        self.image_label.setFixedSize(self.view_size, self.view_size)
+        self.paintingEditor.reset()
 
     def setButtonEnabled(self, value):
         listFull = self.packConrols.updateButtonEnabled()
         self.save_draft_action.setEnabled(listFull)
         self.packConrols.add_button.setEnabled(value)
-        self.view_slider.setEnabled(value)
+        self.paintingEditor.setButtonEnabled(value)
         return
 
     def loadFromFile(self):
@@ -259,44 +96,30 @@ class PaintingStudio(QMainWindow):
         self.packConrols.saveDraft()
 
     def setCurrentImage(self, file_path):
-        # Put loaded image on the file stack
-        self.file_path_stack.append(QUrl(f'file://{file_path}'))
-        self.init_stack_count = len(self.file_path_stack)
-        # Process image
-        self.getNextImage()
+        self.paintingEditor.setCurrentImage(file_path)
 
     def setCurrentData(self, paintingName, paintingMetaData):
-        # Load Meta Data
-        print(paintingMetaData)
-        detail = paintingMetaData["detail"]
-        frameName = paintingMetaData["frameName"]
-        size = paintingMetaData["size"]
-        scale_method = paintingMetaData["scale_method"]
-        background_color = paintingMetaData["background_color"]
-        file_path = paintingMetaData["file_path"]
-        # Set Options
-        if not self.updating:
-            self.updating = True
-            self.detail_spin_box.setValue(detail)
-            self.backgroundColor = background_color
-            self.scale_combo_box.setCurrentText(scale_method)
-            self.size_combo_box.setCurrentText(size)
-            self.painting_combo_box.setCurrentText(paintingName)
-            self.frame_combo_box.setCurrentText(frameName)
-            self.updating = False
+        self.paintingEditor.setCurrentData(paintingName, paintingMetaData)
 
     def getCurrentImageData(self):
-        paintingName = self.painting_combo_box.currentText()
-        imageData = {}
-        imageData[paintingName] = {}
-        imageData[paintingName]['detail']  = self.detail_spin_box.value()
-        imageData[paintingName]['frameName'] = self.frame_combo_box.currentText()
-        imageData[paintingName]['size'] = self.size_combo_box.currentText()
-        imageData[paintingName]['scale_method'] = self.scale_combo_box.currentText()
-        imageData[paintingName]['background_color'] = self.backgroundColor
-        imageData[paintingName]['frame_index'] = self.frame_combo_box.currentIndex()
-        imageData[paintingName]['file_path'] = self.art_path
-        return imageData, paintingName, self.painting
+        return self.paintingEditor.getCurrentImageData()
+
+    def getCurrentImage(self):
+        return self.paintingEditor.getCurrentImage()
+
+    def setLockStatus(self, status):
+        self.paintingEditor.lock = status
+
+    def getNextImage(self):
+        self.setLockStatus(False)
+        self.paintingEditor.getNextImage()
+
+    def updateComboBox(self):
+        self.paintingEditor.updateComboBox()
+
+    def addToComboBox(self, item):
+        if self.paintingEditor.size_combo_box.findText(item) == -1:
+            self.paintingEditor.size_combo_box.addItem(item)
 
     def dragEnterEvent(self, event):
         if event.mimeData().hasUrls():
@@ -306,163 +129,17 @@ class PaintingStudio(QMainWindow):
 
     def dropEvent(self, event):
         # Get the dropped file path
-        if self.packCreated == True:
+        if self.packConrols.packCreated == True:
             for file in event.mimeData().urls():
-                self.file_path_stack.append(file)
-            self.init_stack_count = len(self.file_path_stack)
+                self.paintingEditor.file_path_stack.append(file)
+            self.paintingEditor.init_stack_count = len(self.paintingEditor.file_path_stack)
             self.getNextImage()
         else:
             QMessageBox.information(self, "Pack not Created", f"Please create a pack before importing images.")
 
-    def showColorDialog(self):
-        # Open the QColorDialog
-        color = QColorDialog.getColor(QColor(self.backgroundColor))
-
-        if color.isValid():
-            # Update the label with the chosen color
-            self.backgroundColor = color.name()
-        self.requestViewPortDraw()
-
     def prog_help(self):
-        pass
-
-    def getNextImage(self):
-        if len(self.file_path_stack) > 0:
-            try:
-                url = self.file_path_stack.pop()
-                self.lock = False
-                if url.toLocalFile() == "":
-                    self.art_path = url.toString()
-                    response = requests.get(self.art_path)
-                    img_data = BytesIO(response.content)
-                    self.art = Image.open(img_data)
-                    print(response.status_code)
-                else:
-                    self.art_path = url.toLocalFile()
-                    self.art = Image.open(self.art_path)
-
-                file_name = Path(self.art_path).name.split(".")[0].lower()
-                self.forceViewPortDraw()
-                self.autoSetComboBoxes(file_name)
-                curr = self.init_stack_count-len(self.file_path_stack)
-                self.path_label.setText(f"File: [{curr}/{self.init_stack_count}] - {self.art_path}")
-                self.setButtonEnabled(True)
-            except Exception as e:
-                self.getNextImage()
-                self.image_label.setText(f"Failed to open image: {str(e)}")
-        else:
-            self.init_stack_count = 0
-            #print("Stack is empty.")
-            self.lock = True
-            self.updateComboBox()
-            self.setButtonEnabled(False)
-            self.image_label.clear()
-            self.image_label.setText("Drop Next image here")
-            self.path_label.setText("")
-            self.view_size = 400
-            self.image_label.setFixedSize(self.view_size, self.view_size)
-
-    def requestViewPortDraw(self):
-        if self.drawThread.isActive():
-            #print("WARN: Time delta short. ViewPort Locked")
-            return
-        if self.updating == True:
-            #print("WARN: blocked update, update in progress.")
-            return
-        if self.lock == True:
-            #print("WARN: A push to the image view was preformed while it was locked!")
-            return
-        self.drawThread.start(16) # 16ms frame delta
-
-    def forceViewPortDraw(self):
-        if self.updating == True or self.lock == True:
-            print("WARN: Locked before Delta timeout. Skipping frame.")
-            return
-        #print("Redrawing Viewport.")
-
-        #Setting up UI
-        self.update_view_size()
-
-        # Getting Options
-        detail = self.detail_spin_box.value()
-        scale_method = self.scale_combo_box.currentText()
-        size = self.size_combo_box.currentText()
-        painting = self.painting_combo_box.currentIndex()
-        frame = self.frame_combo_box.currentIndex()
-        if self.frame_combo_box.currentText() == "None":
-            showFrame = False
-            frameName = self.paintings[size][0]
-        else:
-            showFrame = True
-            frameName = self.paintings[size][frame]
-        self.painting = self.painting_maker.makePaiting(detail, scale_method, self.backgroundColor, frameName, showFrame, self.art)
-
-        # Display the image in viewport
-        pil_image = self.painting.convert("RGB")
-        data = pil_image.tobytes("raw", "RGB")
-        qim = QImage(data, pil_image.width, pil_image.height, QImage.Format_RGB888)
-        pixmap = QPixmap(QPixmap.fromImage(qim))
-        self.image_label.setPixmap(pixmap.scaled(QSize(self.view_size, self.view_size), aspectRatioMode=1))
-
-    def autoSetComboBoxes(self, filename):
-        try:
-            if self.updating == True:
-                return
-            options = filename.split("-")
-            paintingName = options[0]
-            for size, painting_list in self.paintings.items():
-                for painting in painting_list:
-                    if painting == paintingName:
-                        self.size_combo_box.setCurrentText(size)
-                        break
-            self.painting_combo_box.setCurrentText(paintingName)
-
-            if len(options) > 1:
-                for option in options:
-                    if option.isdigit():
-                        if int(option) in range(1,17):
-                            self.detail_spin_box.setValue(int(option))
-                    elif option.title() in self.scaleOptions:
-                        self.scale_combo_box.setCurrentText(option.title())
-                self.frame_combo_box.setCurrentText(options[1])
-        except:
-            print("WARN: Failed to parse auto values")
-            pass
-
-
-    def updateFrameComboBox(self):
-        if self.frame_combo_box.currentText() != "None":
-            self.frame_combo_box.setCurrentText(self.painting_combo_box.currentText())
-
-    def updateComboBox(self):
-        self.updating = True
-        size = self.size_combo_box.currentText()
-        self.frame_combo_box.clear()
-        self.painting_combo_box.clear()
-        if size == "":
-            #print("No Paintings.")
-            return
-        for types in self.paintings[size]:
-            self.frame_combo_box.addItem(types)
-            if types not in self.packConrols.used_paintings:
-                self.painting_combo_box.addItem(types)
-        self.frame_combo_box.addItem("None")
-        if self.painting_combo_box.currentText() == "":
-            self.size_combo_box.removeItem(self.size_combo_box.currentIndex())
-        self.updating = False
-        try:
-            self.requestViewPortDraw()
-        except Exception as e:
-            print(f"Failed to open image: {str(e)}")
-        #self.updating = False
-
-
-    def resource_path(self, file):
-        if getattr(sys, 'frozen', False):
-            base_path = sys._MEIPASS
-        else:
-            base_path = os.path.dirname(__file__)
-        return os.path.join(base_path, 'assets', file)
+        dialog = HelpDialog(self)
+        dialog.exec_()
 
 
 def set_theme(app):
